add auto rm ~/.vimrc commad

+++ config.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
-ln -s `pwd`/vimrc ~/.vimrc	
+if [ -f ~/.vimrc ]; then
+	cp ~/.vimrc .vimrc.old
+	rm ~/.vimrc 
+	ln -s `pwd`/vimrc ~/.vimrc	
+fi
